Updated project structure

+++ src/main/java/it/unibo/vampireio/model/Positionable.java
@@ -5,5 +5,5 @@ import java.awt.geom.Point2D;
 public interface Positionable {
     Point2D.Double getPosition();
     void setPosition(Point2D.Double position);
-    int getDistance(Positionable positionable);
+    double getDistance(Positionable positionable);
 }
